feat: faq page updates
