+++ ddp-workspace/projects/ddp-osteo/src/app/app.module.ts
@@ -43,6 +43,7 @@ import { ActivityPageComponent } from './components/activity-page/activity-page.
 import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { UserActivitiesComponent } from './components/user-activities/user-activities.component';
 import { MatTableModule } from '@angular/material/table';
+import { FaqSectionComponent } from './components/faq-section/faq-section.component';
 
 
 const baseElt = document.getElementsByTagName('base');
@@ -175,7 +176,8 @@ export function translateFactory(translate: TranslateService, injector: Injector
     ActivityComponent,
     ActivityPageComponent,
     DashboardComponent,
-    UserActivitiesComponent
+    UserActivitiesComponent,
+    FaqSectionComponent
 
   ],
   providers: [
