+++ ddp-workspace/projects/ddp-osteo/src/app/components/faq-section/faq-section.component.scss
@@ -1,0 +1,85 @@
+@use '@angular/material' as mat;
+@import "../../../app-theme.scss";
+
+
+.faq-section {
+    background-color: mat.get-color-from-palette($app-theme, 350);
+}
+
+.intro-faq-block__content {
+    padding: 2rem 0 0 0;
+}
+
+.faq-block {
+    padding-top: 1rem;
+    padding-bottom: 3rem;
+}
+
+.faq-block__icon {
+    margin: 0 1rem 0 0;
+    line-height: 2.25rem;
+}
+
+.faq-block__question {
+    font-family: 'Source Sans Pro', sans-serif;
+    font-size: 1.375rem;
+    line-height: 2.25rem;
+}
+
+.faq-block__text {
+    font-family: 'Source Sans Pro', sans-serif;
+    font-size: 1.25rem;
+    line-height: 2.25rem;
+    color: mat.get-color-from-palette($app-theme, 700);
+}
+
+
+
+.mat-expansion-panel {
+    border-radius: unset !important;
+    box-shadow: unset !important;
+    border-top: 1px solid mat.get-color-from-palette($app-theme, 650);
+}
+
+.mat-expansion-panel:last-child {
+    border-bottom: 1px solid mat.get-color-from-palette($app-theme, 650);
+}
+
+.mat-expansion-panel-header {
+    padding: 1rem !important;
+    transition: 0.3s;
+}
+
+.mat-content {
+    align-items: flex-start;
+}
+
+::ng-deep .mat-expansion-panel-body {
+    padding: 0 1rem 1rem calc(2rem + 24px) !important;
+}
+
+::ng-deep .Link {
+    color: #3300ff !important;
+    font-family: sans-serif;
+    white-space: normal;
+}
+
+.expand-btn {
+    color: #282828;
+    font-family: 'Source Sans Pro', sans-serif;
+    font-size: 1rem;
+    line-height: 20px;
+    padding: 0.5rem 1rem;
+    text-transform: uppercase;
+    border: 1px solid #000000;
+    border-radius: 10px;
+    background-color: transparent;
+    display: block;
+    position: absolute;
+    transform: translateX(calc(-100% - 30px));
+
+    ::ng-deep.mat-button-wrapper {
+        position: relative;
+        left: -5px;
+    }
+}
